+++ users_managment/src/main/java/com/foodcourt/usersmanagment/application/handler/impl/UserHandler.java
@@ -9,7 +9,6 @@ import com.foodcourt.usersmanagment.infrastructure.configuration.PasswordEncoder
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Component;
-import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
 @Slf4j
