--- users_managment/src/test/java/com/foodcourt/usersmanagment/domain/usecase/CreatorMocks.java
@@ -1,46 +1,0 @@
-package com.foodcourt.usersmanagment.domain.usecase;
-
-import com.foodcourt.usersmanagment.application.dto.request.OwnerRequestDto;
-import com.foodcourt.usersmanagment.domain.model.OwnerModel;
-import com.foodcourt.usersmanagment.infrastructure.out.jpa.entity.UserEntity;
-
-import java.util.Date;
-
-public class CreatorMocks {
-
-    public static OwnerModel createOwnerModel() {
-        return OwnerModel.builder()
-                .name("John Doe")
-                .lastName("Martinez")
-                .dni("123")
-                .phoneNumber("123456789")
-                .password("securePassword")
-                .email("")
-                .birthdayDate(new Date()).build();
-    }
-
-
-    public static OwnerRequestDto createOwnerRequestDto() {
-        return OwnerRequestDto.builder()
-                .name("John Doe")
-                .lastName("Martinez")
-                .dni("123")
-                .phoneNumber("123456789")
-                .password("securePassword")
-                .email("")
-                .birthdayDate(new Date()).build();
-    }
-
-    public static UserEntity createUserEntity() {
-        return UserEntity.builder()
-                .id(1L)
-                .id_rol(2L)
-                .name("John Doe")
-                .lastName("Martinez")
-                .dni("123")
-                .phoneNumber("123456789")
-                .password("securePassword")
-                .email("")
-                .birthdayDate(new Date()).build();
-    }
-}
