--- users_managment/src/test/java/com/foodcourt/usersmanagment/domain/usecase/UserAdapterTest.java
@@ -1,52 +1,0 @@
-package com.foodcourt.usersmanagment.domain.usecase;
-
-import com.foodcourt.usersmanagment.domain.model.OwnerModel;
-import com.foodcourt.usersmanagment.infrastructure.out.jpa.adapter.UserAdapter;
-import com.foodcourt.usersmanagment.infrastructure.out.jpa.entity.UserEntity;
-import com.foodcourt.usersmanagment.infrastructure.out.jpa.mapper.IUserEntityMapper;
-import com.foodcourt.usersmanagment.infrastructure.out.jpa.repository.IUserRepository;
-import org.junit.jupiter.api.BeforeEach;
-import org.junit.jupiter.api.Test;
-import org.mockito.InjectMocks;
-import org.mockito.Mock;
-import org.mockito.MockitoAnnotations;
-
-import static org.junit.jupiter.api.Assertions.assertEquals;
-import static org.mockito.Mockito.*;
-
-public class UserAdapterTest {
-
-    @Mock
-    IUserRepository userRepository;
-
-    @Mock
-    IUserEntityMapper userEntityMapper;
-
-    @InjectMocks
-    UserAdapter userAdapter;
-
-    @BeforeEach
-    void init(){
-        MockitoAnnotations.openMocks(this);
-    }
-
-    @Test
-    void testSaveOwner() {
-        // Aquí puedes implementar el test para guardar un propietario
-        UserEntity userEntity = CreatorMocks.createUserEntity();
-        OwnerModel ownerModel = CreatorMocks.createOwnerModel();
-        // userAdapter.saveOwner(ownerModel);
-        // verify(userRepository, times(1)).save(any(UserEntity.class));
-
-        when(userEntityMapper.toUserEntity(ownerModel)).thenReturn(userEntity);
-        when(userRepository.save(userEntity)).thenReturn(userEntity);
-        when(userEntityMapper.toOwnerModel(userEntity)).thenReturn(ownerModel);
-
-        OwnerModel model = userAdapter.saveOwner(ownerModel);
-
-        assertEquals(ownerModel, model);
-        verify(userEntityMapper, times(1)).toUserEntity(ownerModel);
-        verify(userEntityMapper, times(1)).toOwnerModel(userEntity);
-        verify(userRepository, times(1)).save(userEntity);
-    }
-}
